Adding voyager navigation and can navigate to the movie list screen
