+++ composeApp/src/commonMain/kotlin/me/androidbox/busbymoviesv2/move_list/presentation/screens/MovieListScreen.kt
@@ -2,48 +2,130 @@ package me.androidbox.busbymoviesv2.move_list.presentation.screens
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.lazy.grid.GridCells
 import androidx.compose.foundation.lazy.grid.LazyVerticalGrid
 import androidx.compose.foundation.lazy.grid.items
 import androidx.compose.foundation.lazy.grid.rememberLazyGridState
+import androidx.compose.material.Badge
+import androidx.compose.material.BadgedBox
+import androidx.compose.material.BottomNavigation
+import androidx.compose.material.BottomNavigationDefaults
+import androidx.compose.material.BottomNavigationItem
 import androidx.compose.material.CircularProgressIndicator
+import androidx.compose.material.Icon
 import androidx.compose.material.MaterialTheme
+import androidx.compose.material.Scaffold
+import androidx.compose.material.Text
+import androidx.compose.material.TopAppBar
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableIntStateOf
+import androidx.compose.runtime.saveable.rememberSaveable
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import me.androidbox.busbymoviesv2.move_list.presentation.MovieListItem
+import me.androidbox.busbymoviesv2.move_list.presentation.MovieListAction
 import me.androidbox.busbymoviesv2.move_list.presentation.MovieListState
+import org.jetbrains.compose.resources.stringResource
 import org.jetbrains.compose.ui.tooling.preview.Preview
 
 @Composable
 fun MovieListScreen(
     movieListState: MovieListState,
+    onMovieListAction: (MovieListAction) -> Unit,
     modifier: Modifier = Modifier
 ) {
 
-    Box(modifier = modifier.fillMaxSize(),
-        contentAlignment = Alignment.Center) {
+    var selectedItemIndex by rememberSaveable {
+        mutableIntStateOf(0)
+    }
 
-        if(movieListState.isLoading) {
-            CircularProgressIndicator()
-        }
-        else {
-            LazyVerticalGrid(
-                columns = GridCells.Fixed(count = 2),
-                modifier = Modifier.fillMaxSize(),
-                state = rememberLazyGridState()
+    Scaffold(
+        topBar = {
+            TopAppBar(
+                title = {
+                    Text(stringResource(resource = listOfBottomMovieListNavigationItems[selectedItemIndex].title))
+                }
+            )
+        },
+        content = { paddingValues ->
+            Box(modifier = modifier.fillMaxSize().padding(paddingValues),
+                contentAlignment = Alignment.Center) {
+
+                if(movieListState.isLoading) {
+                    CircularProgressIndicator()
+                }
+                else {
+                    LazyVerticalGrid(
+                        columns = GridCells.Fixed(count = 2),
+                        modifier = Modifier.fillMaxSize(),
+                        state = rememberLazyGridState()
+                    ) {
+                        items(
+                            items = movieListState.movieList,
+                            key = {
+                                it.id
+                            }) {
+                                movieResult ->
+                            MovieListItem(movieResult)
+                        }
+                    }
+                }
+            }
+        /*    MovieListScreen(
+                movieListState = movieListState,
+                modifier = Modifier.padding(paddingValues))*/
+        },
+        bottomBar = {
+            BottomNavigation(
+                elevation = BottomNavigationDefaults.Elevation
             ) {
-                items(
-                    items = movieListState.movieList,
-                    key = {
-                        it.id
-                    }) {
-                        movieResult ->
-                    MovieListItem(movieResult)
+                listOfBottomMovieListNavigationItems.forEachIndexed { index, bottomNavigationItem ->
+                    BottomNavigationItem(
+                        label = {
+                            Text(text = stringResource(resource = bottomNavigationItem.title))
+                        },
+                        selected = selectedItemIndex == index,
+                        onClick = {
+                            /* Don't trigger a new request if the user is on the same category */
+                            if(selectedItemIndex != index) {
+                                /** Just want to load a different movie list i.e. now playing, trending, popular, upcoming */
+                                /** Just want to load a different movie list i.e. now playing, trending, popular, upcoming */
+                                /*movieListViewModel.onLoginAction(
+                                    MovieListAction.OnMovieListNavigationItemClicked(
+                                        listOfBottomMovieListNavigationItems[index].movieCategory
+                                    )
+                                )*/
+                            }
+
+                            selectedItemIndex = index
+                        },
+                        icon = {
+                            BadgedBox(
+                                badge = {
+                                    if(bottomNavigationItem.badgeCount != null) {
+                                        Badge {
+                                            Text(text = bottomNavigationItem.badgeCount.toString())
+                                        }
+                                    }
+                                    else if(bottomNavigationItem.hasExtra){
+                                        Badge()
+                                    }
+                                }
+                            ) {
+                                Icon(
+                                    imageVector = if (selectedItemIndex == index)
+                                        bottomNavigationItem.selectedIcon else bottomNavigationItem.unSelectedIcon,
+                                    contentDescription = stringResource(resource = bottomNavigationItem.title)
+                                )
+                            }
+                        }
+                    )
                 }
             }
         }
-    }
+    )
 }
 
 @Composable
@@ -51,7 +133,8 @@ fun MovieListScreen(
 fun PreviewMovieListScreen() {
     MaterialTheme {
         MovieListScreen(
-            movieListState = MovieListState()
+            movieListState = MovieListState(),
+            onMovieListAction = {}
         )
     }
 }
