+++ composeApp/src/commonMain/kotlin/me/androidbox/busbymoviesv2/navigation/MovieDetailsScreenRoute.kt
@@ -1,0 +1,4 @@
+package me.androidbox.busbymoviesv2.navigation
+
+data object MovieDetailsScreenRoute {
+}
